pyproject.toml
[tool.poetry]
name = "ee-toolset"
version = "0.1.0"
description = "Eternity Engine Toolset"
authors = ["lbedner <LeonardBedner@iheartmedia.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = ">=3.8.6"
bs4 = "^0.0.1"
chromadb = "^0.4.18"
flet = "0.11.0"
gitpython = "^3.1.40"
icecream = "^2.1.3"
langchain = "^0.0.340"
langchainhub = "^0.1.14"
openai = "^1.3.5"
pdfplumber = "^0.10.3"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.1.0"
python-docx = "^1.1.0"
structlog = "^23.2.0"
tiktoken = "^0.5.1"
unstructured = "^0.11.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
pre-commit = "^3.5.0"
tox = "^4.11.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
